# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search for users by name or username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found

  # ── Friends ───────────────────────────────────────────────────────────────────
...
  /conversations/{userId}:
    get:
      operationId: getDirectMessages
      tags: [messages]
      summary: Get messages in a direct conversation
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendDirectMessage
      tags: [messages]
      summary: Send a direct message
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Groups ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        friendCount:
          type: integer
        groupCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        displayName:
          type: string
          minLength: 1
        bio:
          type: string
        age:
          type: integer
          minimum: 13
          maximum: 120
        avatarUrl:
          type: string
        coverUrl:
          type: string
    UserSummary:
      type: object
      required: [id, clerkId, username, displayName]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        friendStatus:
          type: ["string", "null"]
          enum: ["friends", "pending_sent", "pending_received", null]
...
    Message:
      type: object
      required: [id, senderId, content, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: integer
        recipientId:
          type: ["integer", "null"]
        groupId:
          type: ["integer", "null"]
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
          enum: ["image", "video", null]
        sender:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
...